ibus: Check for socket paths involving WAYLAND_DISPLAY first
A recent change in IBus made it prefer the WAYLAND_DISPLAY envvar in
order to compose its socket path for Wayland sessions. This is because
DISPLAY is unreliable in those environment: It might not be there, there
might be several displays pointing to the same Xwayland server (as it's
the case in GNOME 3.36), or there might even be multiple Xwayland servers
(eg. to enforce inter-app isolation with X11 apps).

As the Qt IM does not know which version of IBus it is talking with, make
it prefer the WAYLAND_DISPLAY envvar in its replication of the IBus socket
path. The fallback paths will only if that file does not exist (i.e.
talking to an older IBus, or on a X11 session).