Wasm: don’t deadlock on parallel image conversions
A special restriction of threads on WebAssembly is
that you should not block the main thread, also not
to wait for worker threads.

For example, blocking the main thread may prevent the
browser from starting a new web worker to service the
pthread the main thread is waiting for.

We may be able create an abstraction to support use
cases like this (most likely using emscripten asyncify),
but for disable use of threads to avoid deadlocking.

Change-Id: I35edd5e1bb465e2549fa7cc4288b47dcd2e4244b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>